backend/requirements.txt
fastapi
uvicorn
llama-cpp-python
pydantic
fastapi[all]
asyncio
